Add LazyLoad to Top Table List and Small Promo
Adding client side lazy loading of component and data fetching to top table list and small promo blocks
